Updated to Xcode 9.2 Updated to iOS 10.3 (except for old style popover controller) No more ATS (App Transport Security) exception, use HTTPS from now on Enable data protection for files in entitlements Flickr API key now stored in JSON config file external to project, and copied into app bundle Refreshing from tags screen now clears the cache and database

+++ Core Data SPoT/PhotoImageVC.m
@@ -117,7 +117,7 @@
     self.zooming = YES;
 }
 
-- (void)scrollViewDidEndZooming:(UIScrollView *)scrollView withView:(UIView *)view atScale:(float)scale
+- (void)scrollViewDidEndZooming:(UIScrollView *)scrollView withView:(UIView *)view atScale:(CGFloat)scale
 {
     self.zooming = NO;
 }
